backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
requests
moviepy<2.0
edge-tts
ollama
Pillow<10.0.0
numpy
python-dotenv
apscheduler
duckduckgo_search
diffusers
transformers
accelerate
torch
pydantic-settings
